requirement.txt
Flask>=2.0,<3.0
docxtpl>=0.17.0
python-docx>=0.8.11
waitress>=2.1.2
python-dotenv>=0.21.0
requests